+++ source/extensions/isaacsim.examples.extension/isaacsim/examples/extension/extension.py
@@ -106,9 +106,7 @@ class Extension(omni.ext.IExt):
 
     def _build_info_ui(self):
         title = EXTENSION_NAME
-        doc_link = (
-            "https://docs.omniverse.nvidia.com/isaacsim/latest/advanced_tutorials/tutorial_extension_templates.html"
-        )
+        doc_link = "https://omniverse.gitlab-master-pages.nvidia.com/isaac/isaacsim-docs/utilities/extension_template_generator.html"
 
         overview = (
             "Generate Extension Templates to get started building and programming standalone UI-based extensions in "