+++ source/extensions/isaacsim.robot.wheeled_robots.ui/isaacsim/robot/wheeled_robots/ui/menu_graphs.py
@@ -22,9 +22,7 @@ from omni.kit.notification_manager import NotificationStatus, post_notification
 from omni.kit.window.extensions import SimpleCheckBox
 from pxr import UsdPhysics
 
-OG_DOCS_LINK = (
-    "https://docs.omniverse.nvidia.com/isaacsim/latest/advanced_tutorials/tutorial_advanced_omnigraph_shortcuts.html"
-)
+OG_DOCS_LINK = "https://omniverse.gitlab-master-pages.nvidia.com/isaac/isaacsim-docs/omnigraph/omnigraph_shortcuts.html"
 
 
 class DifferentialRobotGraph: