+++ source/extensions/isaacsim.robot_setup.assembler/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [2.1.2] - 2025-01-17
+### Changed
+- Temporarily changed docs link to gitlab pages
+
+
 ## [2.1.1] - 2024-12-03
 ### Changed
 - Isaac Util menu to Tools->Robotics menu